% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_functions.R
\name{calculate_ld_decay}
\alias{calculate_ld_decay}
\title{Bin LD values for decay plotting}
\usage{
calculate_ld_decay(input, bin_size = 200, max_dist = 50000,
  stat = "r2")
}
\arguments{
\item{input}{A tibble generated by load_plink_ld}

\item{bin_size}{Size of bins, in bp, in which to calculate LD}

\item{max_dist}{Maximum distance between SNPs to allow.}

\item{stat}{Calculate decay for r2 or dp (dprime)?}
}
\value{
A tibble
}
\description{
Bin LD values for decay plotting
}
\examples{
a <- system.file("extdata", "test.plink.ld.gz", package = "gwplotting" )
b <- load_plink_ld( a, max_dist = 10000 )
c <- calculate_ld_decay( b, bin_size = 200 )
c
}
