% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_functions.R
\name{calculate_windowed_ld}
\alias{calculate_windowed_ld}
\title{Calculate average LD values in sliding windows}
\usage{
calculate_windowed_ld(input, window, step)
}
\arguments{
\item{input}{A tibble constructed with load_plink_ld}

\item{window}{Size of window, in bp, to calculate mean LD in}

\item{step}{Step size between windows}
}
\value{
A four-column tibble containing summary statistics
}
\description{
This function will take as input a tibble containing a PLINK LD file
loaded with load_plink_ld(), calculate mean LD values (based on the stat
column values), and return a new tibble.
}
