% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering_functions.R
\name{generate_reordering_for_vcf}
\alias{generate_reordering_for_vcf}
\title{Generate reordering information for a VCF file}
\usage{
generate_reordering_for_vcf(vcf, scaffold_lengths, assignments, species, out)
}
\arguments{
\item{vcf}{The name of the VCF file that you want to reorder.}

\item{scaffold_lengths}{Draft genome scaffold lengths.}

\item{assignments}{Name of the file containing the reordering information.}

\item{species}{The abbreviation for the species assembly you ordered to.}

\item{out}{Prefix of the files to write to.}
}
\value{
Nothing. Produces two files: out.exclude.txt and out.reordering.txt for
    use with PLINK and your original VCF file.
}
\description{
This function will produce two files that you can use in conjunction with
PLINK to reorder your VCF according to chromosome assignments.
}
