% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_functions.R
\name{ld_roller}
\alias{ld_roller}
\title{Function for applying to single-scaffold LD tibbles. Used by the
calculate_windowed_ld function.}
\usage{
ld_roller(input, window, step)
}
\arguments{
\item{input}{A tibble}

\item{window}{Window size, in bp}

\item{step}{Step size, in bp}
}
\value{
A summary tibble
}
\description{
Function for applying to single-scaffold LD tibbles. Used by the
calculate_windowed_ld function.
}
