% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_functions.R
\name{load_abbababa}
\alias{load_abbababa}
\title{Load a results file from Simon Martin's ABBABABAwindows.py script.}
\usage{
load_abbababa(file, stat = "D", position = "midpoint", min_sites = 0)
}
\arguments{
\item{file}{Input file name}

\item{stat}{Statistic(s) to keep. This can be a vector of column names.}

\item{position}{Use the start, end, or midpoint as the position?}

\item{min_sites}{Minimum number of sitesUsed required for window inclusion.}
}
\value{
A tibble.
}
\description{
This function was developed strictly following his GitHub description of the
output file. The columns in this output file are: scaffold, start, end,
sites, sitesUsed, ABBA, BABA, D, fd, fdM. You can choose to plot any of the
statistics, and filter windows by the number of sites used.
}
\examples{
a <- system.file("extdata", "test.abbababa.csv.gz",
                 package = "gwplotting")
b <- load_abbababa( a, stat = 'D' )
b
}
