% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_functions.R
\name{load_pgw}
\alias{load_pgw}
\title{Load a results file from Simon Martin's popgenWindows.py script.}
\usage{
load_pgw(file, position = "midpoint", min_sites = 0)
}
\arguments{
\item{file}{Input file name}

\item{position}{Use the start, end, or midpoint as the position?}

\item{min_sites}{Minimum number of sites required for window inclusion.}
}
\value{
A tibble.
}
\description{
This function was developed strictly following his GitHub description of the
output file. The columns in this output file will vary depending on how many
populations you specify. It is up to you to determine what are the necessary
columns to keep. Minimally, the output from that script is: scaffold, start,
end, mid, sites, pi_popA, pi_popB, dxy_popA_popB, Fst_popA_popB. But you may
have all the pairwise comparisons for each specified populations. This
function will return all of the statistic columns.
}
\examples{
a <- system.file("extdata", "test.popgen.csv.gz",
                 package = "gwplotting")
b <- load_pgw( a )
b
}
