% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_functions.R
\name{load_plink_gwas}
\alias{load_plink_gwas}
\title{Load a PLINK association results file.}
\usage{
load_plink_gwas(file)
}
\arguments{
\item{file}{PLINK association file}
}
\value{
A three-column tibble containing scaffold, bin midpoint, and
    statistic value.
}
\description{
This function will read a (possibly gzipped) PLINK results file (.assoc.txt)
and return a tibble containing just the scaffold, position, and statistic.
}
\examples{
a <- system.file("extdata", "test.plink_gwas.qassoc.gz",
                 package = "gwplotting")
b <- load_plink_gwas( a )
b
}
