% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_functions.R
\name{load_plink_ld}
\alias{load_plink_ld}
\title{Load a plink LD file}
\usage{
load_plink_ld(file, max_dist = 50000, keep_ids = FALSE, stat = c("r2", "dp"))
}
\arguments{
\item{file}{Name of a file containing LD values calculated using}

\item{max_dist}{Maximum allowed distance between two SNPs, in bp}

\item{keep_ids}{Keep SNP IDs in returned tibble? T / F}

\item{stat}{Which statistic to keep? r2, d, and/or dp (D' and D'-signed)}
}
\value{
A tibble with pairwise LD measures
}
\description{
This function loads a file containing pairwise LD estimates generated by
plink. A typical command to generate this file would be:
\emph{plink --bfile input --allow-extra-chr --r2 inter-chr gz \
--ld-window-r2 0.0 --out test.plink --chr scaffold_1000}
This will only keep SNPs on the same chromosome, because it assumes that you
do not know the distance between different scaffolds, even those located to
the same chromosome according to reordering information.
}
\examples{
a <- system.file("extdata", "test.plink.ld.gz", package = "gwplotting" )
b <- load_plink_ld( a )
b
}
