% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loading_functions.R
\name{load_vcftools_stats}
\alias{load_vcftools_stats}
\title{Load a VCFtools results file.}
\usage{
load_vcftools_stats(
  file,
  stat = "mean_fst",
  min_sites = 0,
  position = "midpoint"
)
}
\arguments{
\item{file}{VCFtools results file. Can be gzipped.}

\item{stat}{Name of the column to use as statistic. This will be mean_fst,
weighted_fst, pi, tajimas_d, ...}

\item{min_sites}{Minimum number of sites required to keep a window}

\item{position}{Coordinate to use as position: start, end, or midpoint}
}
\value{
A three-column tibble containing scaffold, bin position, and
    statistic
}
\description{
This function will read a (possibly gzipped) VCFtools results file
from windowed Fst analysis (.windowed.weir.fst), nucleotide diversity
analysis (.windowed.pi), or Tajima's D (.Tajima.D) and return a tibble
containing just the scaffold, position, and statistic.
}
\examples{
a <- system.file("extdata", "test.vcftools_pi.txt.gz",
                 package = "gwplotting")
b <- load_vcftools_stats( a, stat = 'pi' )
b

}
