% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_region_data}
\alias{plot_region_data}
\title{Plot statistics for a particular region}
\usage{
plot_region_data(input, chromosome, start, end, type,
  plotting_column = "stat", scaffold_lengths)
}
\arguments{
\item{input}{A tibble containing}

\item{chromosome}{Chromosome to plot. This will be coded in the "chr"
column of your tibble.}

\item{start}{Start position, if desired}

\item{end}{End position, if desired}

\item{type}{Type of statistic. gwas, popgen. gwas will -log10 the stat, while
popgen will estimate the ylimits.}

\item{plotting_column}{Name of the column to plot. Usually will be 'stat'}

\item{scaffold_lengths}{Name of a two-column tab delimited file with
scaffold and length.}
}
\value{
A ggplot2 object containing the plot.
}
\description{
This is a function to plot statistics on a particular chromosome, or in a
particular region of that chromosome. It takes as input any tibble that was
generated by one of the load functions, the region that you want to plot, and
the statistic and generates a ggplot2 object for printing.
}
\examples{
a1 <- system.file("extdata", "test.gemma_gwas.txt.gz",
                 package = "gwplotting")
a2 <- system.file("extdata", "test.chromSizes.txt.gz",
                  package="gwplotting")

b <- load_gemma_gwas( a1, pval = 'p_wald' )
chrom <- b$chr[ b$scaf == "scaffold_1000" ][1]

c <- plot_region_data( b, chromosome = chrom, type = 'gwas',
plotting_column = 'stat', scaffold_lengths = a2)
c
}
