% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering_functions.R
\name{reorder_by_chromosome}
\alias{reorder_by_chromosome}
\title{Reorder a file by chromosome.}
\usage{
reorder_by_chromosome(input, exclude = T)
}
\arguments{
\item{input}{An input stats file generated using one of the file loading
functions such as load_gemma_gwas}

\item{exclude}{Exclude scaffolds that do not start with "chr" (i.e.
unassigned scaffolds)?}
}
\value{
A tibble
}
\description{
This function will use chromosome information in the input file to order the
results rather than information from an external AGP, etc. file.
}
