% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering_functions.R
\name{reorder_by_scaf_len}
\alias{reorder_by_scaf_len}
\title{Reorder scaffolds based on their size.}
\usage{
reorder_by_scaf_len(input, scaffold_lengths, min_length = 0)
}
\arguments{
\item{input}{A three-column tibble containing (at minimum) scaffold, ps, and
stat as the first three columns.}

\item{scaffold_lengths}{Name of the two-column tab-delimited file containing
scaffold name and length.}

\item{min_length}{Minimum length of scaffolds to keep}
}
\value{
A four-column tibble with scaffolds ordered in increasing size.
}
\description{
This function will order a genome stats tibble according to increasing
scaffold size.
}
\examples{
a1 <- system.file("extdata", "test.gemma_gwas.txt.gz",
                 package = "gwplotting")

a2 <- system.file("extdata", "test.chromSizes.txt.gz",
                  package = "gwplotting" )

b <- load_gemma_gwas( a1, pval = 'p_wald' )
c <- reorder_by_scaf_len( b, a2 )
c
}
