% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering_functions.R
\name{reorder_scaffolds}
\alias{reorder_scaffolds}
\title{Reorder scaffolds based on assignment to chromosomes.}
\usage{
reorder_scaffolds(input, assignments)
}
\arguments{
\item{input}{A three-column tibble containing (at minimum) scaffold, ps, and
stat as the first three columns.}

\item{assignments}{Name of the file containing the reordering information.}
}
\value{
Returns the input tibble with a new column, chr, detailing which
    reference genome chromosomes the draft genome scaffolds map to. Only
    input rows that reside on mapped scaffolds are included.
}
\description{
This function will reorder the results of a genome-wide statistical
analysis based on the contents of a file that assigns each scaffold to a
reference genome assembly. The assignments can be generated using, e.g.
OrderScaffoldsByBlatingProteins.pl, but must be in the following format:
}
\details{
The scaffold_info file can be generated using UCSCtools' faToTwoBit and
twoBitInfo.
}
\examples{
a1 <- system.file("extdata", "test.gemma_gwas.txt.gz",
                  package = "gwplotting")

a2 <- system.file("extdata", "test.reordering.txt.gz",
                  package = "gwplotting")

b <- load_gemma_gwas( a1, pval = 'p_wald' )
b <- reorder_scaffolds( b, a2 )
b
}
