% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_region_data}
\alias{plot_region_data}
\title{Plot statistics in a particular region}
\usage{
plot_region_data(
  input,
  chromosome = NA,
  scaffold = NA,
  from = 1,
  to = NA,
  type = "gwas",
  color = "black",
  plotting_column = "stat",
  scaffold_lengths,
  include_all_p = FALSE
)
}
\arguments{
\item{input}{A tibble with four named columns (scaf, ps, stat, and chr)}

\item{chromosome}{Chromosome to plot. This OR scaffold may be specified.}

\item{scaffold}{Scaffold to plot. This OR chromosome may be specified.}

\item{from}{Start position, if desired.}

\item{to}{End position, if desired}

\item{type}{Type of statistic: gwas, popgen.}

\item{color}{Color for points.}

\item{plotting_column}{Name of the column to plot. Usually will be 'stat'}

\item{scaffold_lengths}{Path to a two-column tab delimited file with
scaffold and length.}

\item{include_all_p}{Plot p-values <0.05?}
}
\value{
A ggplot.
}
\description{
This is a function to plot statistics on a particular scaffold or chromosome,
or in a particular region of that scaffold or chromosome. The input may be 1)
the raw tibble from any of the loading functions, 2) a reordered tibble, or
3) a tibble with cumulative positions. This function will operate differently
on each of these input types. If the input already has cumulative positions
(i.e. column bp_cum), then the "from" and "to" variables are interpreted as
"from = bp_cum start" and "to = bp_cum end". If the input does not have
column bp_cum, then this function will generate bp_cum according to the
desired chromosome or scaffold, assume that the supplied "from" and "to"
arguments are scaffold positions, and determine the appropriate "from" and
"to" coordinates from the new bp_cum coordinates. You may only specify a
chromosome OR a scaffold.
}
\examples{
a1 <- system.file("extdata", "test.gemma_gwas.txt.gz",
                 package = "gwplotting")
a2 <- system.file("extdata", "test.chromSizes.txt.gz",
                  package="gwplotting")

b <- load_gemma_gwas( a1, pval = 'p_wald' )
chrom <- b$chr[ b$scaf == "scaffold_1000" ][1]

c <- plot_region_data( b, chromosome = chrom, type = 'gwas',
plotting_column = 'stat', scaffold_lengths = a2)
c
}
